Port most snippets from old config
Note that whenever `snipMate` snippets (in `snippets/*.snippets`) use
the `VISUAL` placeholder in a non-trivial form, it will not be replaced
by the actual selection but stay as-is. The only form where it works is
in as `${VISUAL}` and not as `$VISUAL` or even inside other placeholder
default values.

Note also the text substitutions don't work.

Both of the above issues can be resolved by porting the snippets to Lua.
The `utils.luasnip_utils.dyn_sel()` function simulates
`${1:${VISUAL:...}}$` (with configurable index and default string). See
`lua/snippets/python.lua` for examples.